+++ app/response/utils.py
@@ -56,7 +56,7 @@ from app.lib.date_utils import (
 from app.lib.db_utils import create_object, update_object, delete_object
 from app.lib.email_utils import send_email, get_agency_emails
 from app.lib.redis_utils import redis_get_file_metadata, redis_delete_file_metadata
-from app.lib.utils import eval_request_bool, UserRequestException
+from app.lib.utils import eval_request_bool, UserRequestException, DuplicateFileException
 from app.models import (
     Events,
     Notes,
@@ -82,12 +82,10 @@ def add_file(request_id, filename, title, privacy, is_editable):
     Gets the file mimetype and magic file check from a helper function in lib.file_utils
     File privacy options can be either Release and Public, Release and Private, or Private.
     Provides parameters for the process_response function to create and store responses and events object.
-
     :param request_id: Request ID that the file is being added to
     :param filename: The secured_filename of the file.
     :param title: The title of the file which is entered by the uploader.
     :param privacy: The privacy option of the file.
-
     """
     path = os.path.join(current_app.config['UPLOAD_DIRECTORY'], request_id, filename)
     try:
@@ -98,21 +96,24 @@ def add_file(request_id, filename, title, privacy, is_editable):
         mime_type = fu.get_mime_type(path)
         hash_ = fu.get_hash(path)
 
-    response = Files(
-        request_id,
-        privacy,
-        title,
-        filename,
-        mime_type,
-        size,
-        hash_,
-        is_editable=is_editable
-    )
-    create_object(response)
+    try:
+        response = Files(
+            request_id,
+            privacy,
+            title,
+            filename,
+            mime_type,
+            size,
+            hash_,
+            is_editable=is_editable
+        )
+        create_object(response)
 
-    create_response_event(event_type.FILE_ADDED, response)
+        create_response_event(event_type.FILE_ADDED, response)
 
-    return response
+        return response
+    except DuplicateFileException as e:
+        return str(e)
 
 
 def add_note(request_id, note_content, email_content, privacy, is_editable, is_requester):
