+++ app/response/views.py
@@ -136,7 +136,10 @@ def response_file(request_id):
                                 files[file_data]['title'],
                                 files[file_data]['privacy'],
                                 is_editable=True)
-        get_file_links(response_obj, release_public_links, release_private_links, private_links)
+        if not isinstance(response_obj, Files):
+            flash(message=response_obj, category='danger')
+        else:
+            get_file_links(response_obj, release_public_links, release_private_links, private_links)
     send_file_email(request_id,
                     release_public_links,
                     release_private_links,
